GOF for circles
Added a test for goodness-of-fit for circle fit after clustering. The
GOF value after removal of two weakest clusters is not used as
additional cut-off, with default of 0.98 (can be increased to 0.99).
The clustered images and GOF values are now saved in a separate folder
in results. Finally, after data is clustered the program ask the user
if clustering is ok, if yes - it saves the images, values and proceed
to the next ring, if not - attempt to recluster.